修改自定义JumpAdapter设置方式，增加权限校验

+++ windowtree_annotation/src/main/java/com/kaixuan/windowtree_annotation/annotation/Window.java
@@ -40,5 +40,5 @@ public @interface Window {
      * 该值默认为-1时，即该页面无权限控制，任何人都允许加载
      * @return
      */
-    int pageAuthority() default -1;
+    long pageAuthority() default -1;
 }
